# Generalize `vec_split_along()` to implement `vec_split()` in C