--- a/src/utils.h
+++ b/src/utils.h
@@ -216,6 +216,7 @@ extern SEXP strings_universal;
 extern SEXP strings_check_unique;
 extern SEXP strings_key;
 extern SEXP strings_id;
+extern SEXP strings_val;
 
 extern SEXP syms_i;
 extern SEXP syms_n;