I've generalized `vec_split_along()` to be more useful. Rather than just mimicking a faster version of:

```
indices <- as.list(vec_seq_along(x))
map(indices, vec_slice, x = x)
```

It can now take an arbitrary list of integer `indices`. The signature is now `vec_split_along(x, indices = NULL)` with `NULL` defaulting to the original behavior. 

The reason this is useful is for using it in `vec_split()`. Currently there is a call to:

```
x_split <- map(out$id, vec_slice, x = x)
```

This can be completely replaced with the new version of `vec_split_along()`, and then a C version of `vec_split()` can be written.

Below are a large number of benchmarks. The tl;dr is that the new versions of `vec_split_along()` and `vec_split()` perform much better than their older alternatives in most cases, and never perform worse. Vectors have a much more dramatic performance boost than data frames. Generally, the more unique values you have to split by, the bigger the performance gap.

`vec_split_along()` benchmarks:

<details>

``` r
library(nycflights13)
library(purrr)
library(vctrs)

vec_split_along <- vctrs:::vec_split_along

x <- flights[1:3]

# `vec_split_along()` generally outperforms repeated `vec_slice()`ing
idx <- rep_len(list(1:50), 10000)
bench::mark(
  vec_split_along(x, idx),
  map(idx, vec_slice, x = x),
  iterations = 100
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_along(x, idx)      12.4ms   23.5ms     35.7     7.17MB     14.3
#> 2 map(idx, vec_slice, x = x)   66.2ms  104.6ms      9.20    7.22MB     24.1

# take a long factor and repeatedly subset the first 1000 values
z <- factor(rep(letters, times = 1000))
z_idx <- rep_len(list(1:1000), 1000)

bench::mark(
  vec_split_along(z, z_idx),
  map(z_idx, vec_slice, x = z)
)
#> # A tibble: 2 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_along(z, z_idx)      9.13ms 10.7ms      93.0    3.88MB     22.6
#> 2 map(z_idx, vec_slice, x = z)  13.85ms   15ms      62.3    3.87MB     19.8

# take a long integer and repeatedly subset the first 1000 values
w <- 1:100000 + 0L
w_idx <- rep_len(list(1:1000), 1000)

bench::mark(
  vec_split_along(w, w_idx),
  map(w_idx, vec_slice, x = w),
  iterations = 1000
)
#> # A tibble: 2 x 6
#>   expression                       min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_along(w, w_idx)     1.32ms  1.74ms      544.    3.87MB     32.9
#> 2 map(w_idx, vec_slice, x = w)  1.95ms  2.37ms      409.    3.87MB     22.9

# mimic the defaults of `vec_split_along()`, i.e. split `a` along its size
# (as expected, the default method is slightly faster than supplying the index)
# (but both are much faster than repeated slicing)
a <- 1:100000 + 0L
a_idx <- as.list(vec_seq_along(a))

bench::mark(
  vec_split_along(a),
  vec_split_along(a, a_idx),
  map(a_idx, vec_slice, x = a),
  iterations = 100
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 3 x 6
#>   expression                       min   median `itr/sec` mem_alloc
#>   <bch:expr>                   <bch:t> <bch:tm>     <dbl> <bch:byt>
#> 1 vec_split_along(a)            3.79ms   4.49ms    172.       781KB
#> 2 vec_split_along(a, a_idx)     6.59ms   8.51ms    110.       781KB
#> 3 map(a_idx, vec_slice, x = a) 82.01ms 130.72ms      7.38     781KB
#> # … with 1 more variable: `gc/sec` <dbl>
```

<sup>Created on 2019-07-31 by the [reprex package](https://reprex.tidyverse.org) (v0.2.1)</sup>

</details>

`vec_split()` benchmarks:

<details>

``` r

library(nycflights13)
library(purrr)
library(vctrs)

# A pure C version of vec_split() using the new vec_split_along()
vec_split2 <- vctrs:::vec_split2

# all unique
int <- 1:5000
bench::mark(
  vec_split(int, int),
  vec_split2(int, int)
)
#> # A tibble: 2 x 6
#>   expression                min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>           <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(int, int)    3.81ms    4.6ms      217.     512KB     26.0
#> 2 vec_split2(int, int) 535.31µs  674.7µs     1447.     315KB     21.7

# all same
same_idx <- rep(1L, 5000)
bench::mark(
  vec_split(int, same_idx),
  vec_split2(int, same_idx)
)
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(int, same_idx)     103µs  136.2µs     6916.     174KB     20.6
#> 2 vec_split2(int, same_idx)     79µs   94.9µs    10127.     174KB     33.8

# large all unique
bench::mark(
  vec_split(1:50000, 1:50000),
  vec_split2(1:50000, 1:50000),
  iterations = 50
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:tm> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(1:50000, 1:50000)   57.92ms   85ms      10.6    3.99MB     22.7
#> 2 vec_split2(1:50000, 1:50000)   8.63ms 14.5ms      55.4    3.23MB     17.7

# two groups
idx <- c(rep(1, 2500), rep(2, 2500))
bench::mark(
  vec_split(1:5000, idx),
  vec_split2(1:5000, idx)
)
#> # A tibble: 2 x 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(1:5000, idx)     154µs    192µs     5222.     194KB     18.4
#> 2 vec_split2(1:5000, idx)    123µs    153µs     6594.     194KB     25.0


fctr <- factor(rep(letters, 20000))
length(fctr)
#> [1] 520000

# factor - 2 groups
fctr_idx <- sample.int(2, size = length(fctr), replace = TRUE)
bench::mark(
  vec_split(fctr, fctr_idx),
  vec_split2(fctr, fctr_idx),
  iterations = 100
)
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(fctr, fctr_idx)    11.2ms   12.5ms      80.6    19.9MB     157.
#> 2 vec_split2(fctr, fctr_idx)   10.6ms   11.3ms      87.0    19.9MB     120.

# factor - 1000 unique groups
fctr_idx <- sample.int(1000, size = length(fctr), replace = TRUE)
bench::mark(
  vec_split(fctr, fctr_idx),
  vec_split2(fctr, fctr_idx),
  iterations = 100
)
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(fctr, fctr_idx)    21.6ms   25.1ms      39.6    20.1MB     80.2
#> 2 vec_split2(fctr, fctr_idx)   17.1ms   19.6ms      51.0    20.1MB     65.7


# lists - all unique
lst <- rep_len(list(1), 10000)
lst_idx <- vec_seq_along(lst)
bench::mark(
  vec_split(lst, lst_idx),
  vec_split2(lst, lst_idx)
)
#> # A tibble: 2 x 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(lst, lst_idx)    8.65ms   9.16ms      108.     818KB     38.9
#> 2 vec_split2(lst, lst_idx)   1.12ms   1.37ms      709.     622KB     30.7

# small data frame
bench::mark(
  vec_split(mtcars, mtcars["cyl"]),
  vec_split2(mtcars, mtcars["cyl"])
)
#> # A tibble: 2 x 6
#>   expression                           min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_split(mtcars, mtcars["cyl"])  73.2µs 89.6µs    11172.   18.07KB
#> 2 vec_split2(mtcars, mtcars["cyl"]) 37.7µs 47.8µs    20963.    1.26KB
#> # … with 1 more variable: `gc/sec` <dbl>

# large data frame - split by data frame with few unique values
# (not much performance boost)
bench::mark(
  vec_split(flights, flights["month"]),
  vec_split2(flights, flights["month"]),
  iterations = 100
)
#> # A tibble: 2 x 6
#>   expression                               min median `itr/sec` mem_alloc
#>   <bch:expr>                            <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_split(flights, flights["month"])  39.9ms 47.3ms      21.0   179.5MB
#> 2 vec_split2(flights, flights["month"]) 40.6ms 43.9ms      22.3    48.4MB
#> # … with 1 more variable: `gc/sec` <dbl>

# large data frame - split by vector with few unique values
# (not much performance boost)
bench::mark(
  vec_split(flights, flights$month),
  vec_split2(flights, flights$month),
  iterations = 100
)
#> # A tibble: 2 x 6
#>   expression                            min median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_split(flights, flights$month)  11.7ms 14.1ms      68.7    48.4MB
#> 2 vec_split2(flights, flights$month) 10.9ms 13.6ms      72.4    48.4MB
#> # … with 1 more variable: `gc/sec` <dbl>

# large data frame - split by data frame with many unique values
# (more performance boost here)
vec_unique_count(flights["tailnum"])
#> [1] 4044

bench::mark(
  vec_split(flights, flights["tailnum"]),
  vec_split2(flights, flights["tailnum"]),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 x 6
#>   expression                                min median `itr/sec` mem_alloc
#>   <bch:expr>                              <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split(flights, flights["tailnum"])  215ms  239ms      3.86    51.5MB
#> 2 vec_split2(flights, flights["tailnum"]) 195ms  218ms      4.36    51.4MB
#> # … with 1 more variable: `gc/sec` <dbl>

# large data frame - split by vector with many unique values
# (more performance boost here)
bench::mark(
  vec_split(flights, flights$tailnum),
  vec_split2(flights, flights$tailnum),
  iterations = 10
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 2 x 6
#>   expression                             min median `itr/sec` mem_alloc
#>   <bch:expr>                           <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split(flights, flights$tailnum)  172ms  199ms      4.78    51.5MB
#> 2 vec_split2(flights, flights$tailnum) 137ms  165ms      5.34    51.4MB
#> # … with 1 more variable: `gc/sec` <dbl>

# medium sized data frame
# has dbl/int/chr cols
mini_flights <- flights[1:1000, 6:10]

# split by a vector that only has 14 groups
bench::mark(
  vec_split(mini_flights, mini_flights$carrier),
  vec_split2(mini_flights, mini_flights$carrier)
)
#> # A tibble: 2 x 6
#>   expression                                         min  median `itr/sec`
#>   <bch:expr>                                     <bch:t> <bch:t>     <dbl>
#> 1 vec_split(mini_flights, mini_flights$carrier)  165.8µs 194.1µs     4865.
#> 2 vec_split2(mini_flights, mini_flights$carrier)  60.2µs  78.8µs    12054.
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

# ~2/3 unique
vec_unique_count(mini_flights$arr_time)
#> [1] 640

bench::mark(
  vec_split(mini_flights, mini_flights$arr_time),
  vec_split2(mini_flights, mini_flights$arr_time)
)
#> # A tibble: 2 x 6
#>   expression                                           min median `itr/sec`
#>   <bch:expr>                                      <bch:tm> <bch:>     <dbl>
#> 1 vec_split(mini_flights, mini_flights$arr_time)    4.33ms 4.92ms      188.
#> 2 vec_split2(mini_flights, mini_flights$arr_time)  868.2µs 1.03ms      894.
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

# arrays - all unique
arr <- array(1:20000, c(2000, 10))
arr_idx <- vec_seq_along(arr)
bench::mark(
  vec_split(arr, arr_idx),
  vec_split2(arr, arr_idx)
)
#> # A tibble: 2 x 6
#>   expression                    min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>               <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(arr, arr_idx)    2.75ms   3.12ms      299.     174KB     13.0
#> 2 vec_split2(arr, arr_idx)   1.04ms   1.25ms      738.     134KB     13.2

# arrays - 3 groups
arr2 <- array(1:20000, c(2000, 10))
arr2_idx <- rep(1:3, length.out = nrow(arr2))
bench::mark(
  vec_split(arr2, arr2_idx),
  vec_split2(arr2, arr2_idx)
)
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(arr2, arr2_idx)    91.8µs  129.7µs     7766.     150KB     6.66
#> 2 vec_split2(arr2, arr2_idx)   58.8µs   82.3µs    11246.     150KB     9.37

# arrays - 3 groups - many columns
arr3 <- array(1:20000, c(10, 2000))
arr3_idx <- rep(1:3, length.out = nrow(arr3))
bench::mark(
  vec_split(arr3, arr3_idx),
  vec_split2(arr3, arr3_idx)
)
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split(arr3, arr3_idx)      67µs   90.9µs    10606.    78.3KB     4.22
#> 2 vec_split2(arr3, arr3_idx)   44.4µs   58.2µs    16261.    78.3KB     9.10
```

<sup>Created on 2019-07-31 by the [reprex package](https://reprex.tidyverse.org) (v0.2.1)</sup>

</details>