@DavisVaughan 2019-08-07
I've updated `vec_split_along_df()` to no longer use `df_slice()`. This was inefficient because for the `n` `indices` you would be proxying each column of the data frame `n` times. The new method inverts this. It looks at a single column of `x` at a time, proxy-ing once and performing a complete split of that single column with all `n` indices at once (using another call to `vec_split_along()`). This seems to be more efficient. Below I will list some benchmarks just to have the change from the old method documented.

``` r
library(nycflights13)
library(vctrs)

vec_split_along <- vctrs:::vec_split_along

mini_flights <- flights[1:20000,]
```

```r
# before
bench::mark(vec_split_along(mini_flights), iterations = 200)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                    <bch:t> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_along(mini_flights)   259ms  429ms      2.38    3.97MB     13.6

# after
bench::mark(vec_split_along(mini_flights), iterations = 200)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                        min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                    <bch:t> <bch:>     <dbl> <bch:byt>    <dbl>
#> 1 vec_split_along(mini_flights)   132ms  212ms      4.15    6.87MB     10.9
```

```r
mini_flights_no_POSIXct <- mini_flights[,-19]

# before
bench::mark(vec_split_along(mini_flights_no_POSIXct), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                                  min median `itr/sec`
#>   <bch:expr>                               <bch:> <bch:>     <dbl>
#> 1 vec_split_along(mini_flights_no_POSIXct) 58.4ms 66.9ms      12.3
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>

# after
bench::mark(vec_split_along(mini_flights_no_POSIXct), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                                  min median `itr/sec`
#>   <bch:expr>                               <bch:> <bch:>     <dbl>
#> 1 vec_split_along(mini_flights_no_POSIXct) 29.5ms 37.9ms      25.2
#> # … with 2 more variables: mem_alloc <bch:byt>, `gc/sec` <dbl>
```

```r
a_POSIXlt_column <- mini_flights[,integer()]
a_POSIXlt_column$posixlt <- as.POSIXlt(new_datetime(0))

# before
bench::mark(vec_split_along(a_POSIXlt_column), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(a_POSIXlt_column) 478ms  696ms      1.47     194KB
#> # … with 1 more variable: `gc/sec` <dbl>

# after
bench::mark(vec_split_along(a_POSIXlt_column), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(a_POSIXlt_column) 403ms  597ms      1.76    2.02MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

```r
# split by a vector that has >1000 groups
id <- vec_split_id(mini_flights$tailnum)$id

# before
bench::mark(vec_split_along(mini_flights, id), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                           min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(mini_flights, id) 25.3ms 41.2ms      23.7    1.41MB
#> # … with 1 more variable: `gc/sec` <dbl>

# after
bench::mark(vec_split_along(mini_flights, id), iterations = 200)
#> # A tibble: 1 x 6
#>   expression                           min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch:> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(mini_flights, id) 16.9ms 22.5ms      43.2    2.31MB
#> # … with 1 more variable: `gc/sec` <dbl>
```

```r
# same thing, but with a really wide data frame
wide_flights <- vec_cbind(!!!rep(list(mini_flights), 6))
#> New names:
#> * year -> year...1
#> * month -> month...2
#> * day -> day...3
#> * dep_time -> dep_time...4
#> * sched_dep_time -> sched_dep_time...5
#> * ...
id <- vec_split_id(wide_flights$tailnum...12)$id

# before
bench::mark(vec_split_along(wide_flights, id), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(wide_flights, id) 184ms  275ms      3.47    7.52MB
#> # … with 1 more variable: `gc/sec` <dbl>

# after
bench::mark(vec_split_along(wide_flights, id), iterations = 100)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.
#> # A tibble: 1 x 6
#>   expression                          min median `itr/sec` mem_alloc
#>   <bch:expr>                        <bch> <bch:>     <dbl> <bch:byt>
#> 1 vec_split_along(wide_flights, id) 160ms  258ms      3.39    12.8MB
#> # … with 1 more variable: `gc/sec` <dbl>
```